Integration tests should work for Mac OS.

+++ run_integration_tests.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2022 The Turbo Cache Authors. All rights reserved.
+# Copyright 2023 The Turbo Cache Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,6 +14,11 @@
 # limitations under the License.
 
 set -euo pipefail
+
+if [[ $EUID -eq 0 ]]; then
+  echo "This script should not be run as root."
+  exit 1
+fi
 
 TEST_PATTERNS=()
 
@@ -60,13 +65,24 @@ sudo rm -rf ~/.cache/turbo-cache
 mkdir -p ~/.cache/turbo-cache
 
 # Ensure our docker compose is not running.
-docker-compose rm --stop -f
+sudo docker-compose rm --stop -f
 
 export TMPDIR=$HOME/.cache/turbo-cache/
 mkdir -p "$TMPDIR"
-export CACHE_DIR=$(mktemp -d --suffix="-turbo-cache-integration-test")
+
+if [[ "${OSTYPE:-linux-gnu}" == "linux-gnu"* ]]; then
+  export CACHE_DIR=$(mktemp -d --tmpdir="$TMPDIR" --suffix="-turbo-cache-integration-test")
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+  # Create a temporary directory using mktemp with a random template, then add a suffix.
+  export CACHE_DIR=$(mktemp -d "${TMPDIR}turbo-cache-integration-test")
+else
+  # Unknown OS
+  echo "Unable to detect operating system. Assuming the Linux/WSL syntax will work."
+  export CACHE_DIR=$(mktemp -d --tmpdir="$TMPDIR" --suffix="-turbo-cache-integration-test")
+fi
+
 export BAZEL_CACHE_DIR="$CACHE_DIR/bazel"
-trap "sudo rm -rf $CACHE_DIR; docker-compose rm --stop -f" EXIT
+trap "sudo rm -rf $CACHE_DIR; sudo docker-compose rm --stop -f" EXIT
 
 echo "" # New line.
 
@@ -78,11 +94,18 @@ mkdir -p "$TURBO_CACHE_DIR"
 for pattern in "${TEST_PATTERNS[@]}"; do
   find "$SELF_DIR/integration_tests/" -name "$pattern" -type f -print0 | while IFS= read -r -d $'\0' fullpath; do
     # Cleanup.
-    sudo find "$TURBO_CACHE_DIR" -delete
+    echo "Cleaning up cache directories TURBOC_CACHE_DIR: $TURBO_CACHE_DIR"
+    echo "Checking for existince of the TURBO_CACHE_DIR"
+    if [ -d "$TURBO_CACHE_DIR" ]; then
+      sudo find "$TURBO_CACHE_DIR" -delete # add for linux
+    else
+      echo "Directory $TURBO_CACHE_DIR does not exist."
+    fi
+
     bazel --output_base="$BAZEL_CACHE_DIR" clean
     FILENAME=$(basename $fullpath)
     echo "Running test $FILENAME"
-    docker-compose up -d
+    sudo docker-compose up -d
     set +e
     bash -euo pipefail "$fullpath"
     EXIT_CODE="$?"
@@ -91,10 +114,10 @@ for pattern in "${TEST_PATTERNS[@]}"; do
       echo "$FILENAME passed"
     else
       echo "$FILENAME failed with exit code $EXIT_CODE"
-      docker-compose logs
+      sudo docker-compose logs
       exit $EXIT_CODE
     fi
-    docker-compose rm --stop -f
+    sudo docker-compose rm --stop -f
     echo "" # New line.
   done
 done
